Modify `createVNode` argument types to match `VNode`.
In the previous sample code, the type of the argument `children` of `createVNode` is `unknown`,

> vnode.ts に createVNode と言う関数を作っておいて、h 関数からはそれを呼ぶようにしておきましょう。
>
> ```ts
> export function createVNode(
>   type: VNodeTypes,
>   props: VNodeProps | null,
>   children: unknown
> ): VNode {
>   const vnode: VNode = { type, props, children };
>   return vnode;
> }
> ```
> https://ubugeeei.github.io/chibivue/10-minimum-example/040-minimum-virtual-dom.html#patch-%E9%96%A2%E6%95%B0%E3%81%AE%E5%AE%9F%E8%A3%85%E3%82%92%E8%A1%8C%E3%81%86%E5%89%8D%E3%81%AB

and changing the type of `children` in the `VNode` to `VNodeNormalizedChildren` will result in an error because there is no explicit type cast.

> ```ts
> export const Text = Symbol();
>
> export type VNodeTypes = string | typeof Text;
>
> export interface VNode<HostNode = any> {
>   type: VNodeTypes;
>   props: VNodeProps | null;
>   children: VNodeNormalizedChildren;
> }
> ```

Therefore, I added code to change the type of the argument `children` of `createVNode` to `VNodeNormalizedChildren` in the sample code.