fix: 📌 Added the implementation of advanceSpaces to the sample code in the document.
There is a place where `advanceSpaces` is called in the implementation of `parseTag`,
but since the implementation of `advanceSpaces` did not exist in the document,
I added the implementation of `advanceSpaces` to the utilities in the sample code of the `parseText` implementation.

```ts
function parseTag(context: ParserContext, type: TagType): ElementNode {
  // Tag open.
  const start = getCursor(context)
  const match = /^<\/?([a-z][^\t\r\n\f />]*)/i.exec(context.source)!
  const tag = match[1]

  advanceBy(context, match[0].length)
  advanceSpaces(context)
```

ref: https://ubugeeei.github.io/chibivue/10-minimum-example/070-more-complex-parser.html#parseelement

I referred to the following description for the implementation of advanceSpaces.

https://github.com/Ubugeeei/chibivue/blob/v0.0.8/book/impls/10_minimum_example/060_template_compiler2/packages/compiler-core/parse.ts#L112-L117